Transformers: Extract helper classes to separate file
